requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
PyMuPDF>=1.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
transformers>=4.40.0
torch>=2.2.0
numpy>=1.26.0
